+++ scripts/lib/sockets.js
@@ -90,23 +90,23 @@ async function deleteEmbeddedDocuments(entityUuid, type, ids, options) {
 async function updateTargets(targetIds) {
     canvas.tokens?.setTargets(targetIds);
 }
-async function addFavorites(actorUuid, entityUuids, type='item') {
+async function addFavorites(actorUuid, entityUuids) {
     let actor = await fromUuid(actorUuid);
-    if (!actor) return;
-    if (!actor.system.addFavorite) return;
-    let entities = await Promise.all(entityUuids.map(async i => {
-        return await fromUuid(i);
-    }).filter(j => j));
-    if (type === 'item') {
-        for (let i of entities) await actor.system.addFavorite({
-            id: i.getRelativeUUID(i.actor),
-            type: 'item'
-        });
-    } else if (type === 'activity') {
-        for (let i of entities) await actor.system.addFavorite({
-            id: i.relativeUUID,
-            type: 'activity'
-        });
+    if (!actor || !actor.system.addFavorite) return;
+    let entities = (await Promise.all(entityUuids.map(i => fromUuid(i)))).filter(Boolean);
+    for (let entity of entities) {
+        let entityType = entity.documentName;
+        if (entityType === 'Item') {
+            await actor.system.addFavorite({
+                id: entity.getRelativeUUID(entity.actor),
+                type: 'item'
+            });
+        } else if (entityType === 'Activity') {
+            await actor.system.addFavorite({
+                id: entity.relativeUUID,
+                type: 'activity'
+            });
+        }
     }
 }
 async function removeFavorites(actorUuid, entityUuids, type='item') {
@@ -208,7 +208,13 @@ async function syntheticItemDataRoll(itemData, actorUuid, targetUuids, {options 
     let targets = targetUuids.map(i => fromUuidSync(i)?.object);
     let item = await itemUtils.syntheticItem(itemData, actor);
     let workflow = await workflowUtils.syntheticItemRoll(item, targets, {options, config});
-    return workflow.getMacroData({noWorkflowReference: true});
+    return workflow.getSafeMacroData();
+}
+async function syntheticActivityDataRoll(activityData, itemUuid, targetUuids, {options = {}, config = {}, atLevel = undefined, consumeUsage = false, consumeResources = false} = {}) {
+    let item = fromUuidSync(itemUuid);
+    let targets = targetUuids.map(i => fromUuidSync(i)?.object);
+    let workflow = await workflowUtils.syntheticActivityDataRoll(activityData, item, item.parent, targets, {options, config, atLevel, consumeUsage, consumeResources});
+    return workflow.getSafeMacroData();
 }
 const dialogManager = new DialogManager;
 async function queuedDialog(dialogOptions, checkOptions) {
@@ -257,6 +263,7 @@ export let sockets = {
     remoteRoll,
     remoteDamageRolls,
     syntheticItemDataRoll,
+    syntheticActivityDataRoll,
     queuedDialog
 };
 export let socket;