+++ scripts/lib/utilities/workflowUtils.js
@@ -38,6 +38,9 @@ async function completeActivityUse(activity, config = {}, dialog = {}, message =
         config.midiOptions.checkGMStatus = true;
         config.midiOptions.workflowData = true;
         fixSets = true;
+    } else if (config.midiOptions?.asUser && config.midiOptions?.asUser !== game.userId) {
+        config.midiOptions.workflowData = true;
+        fixSets = true;
     }
     let workflow = await MidiQOL.completeActivityUse(activity, config, dialog, message);
     workflow = workflow.workflow ?? workflow;
@@ -68,7 +71,7 @@ async function completeItemUse(item, config = {}, options = {}) {
     }
     return workflow;
 }
-async function syntheticActivityRoll(activity, targets = [], {options = {}, config = {}, atLevel = undefined, consumeUsage = false, consumeResources = false, spellSlot = false, dialog = {}, message = {}} = {}) {
+async function syntheticActivityRoll(activity, targets = [], {options = {}, config = {}, userId, atLevel = undefined, consumeUsage = false, consumeResources = false, spellSlot = false, dialog = {}, message = {}} = {}) {
     let defaultConfig = {
         consumeUsage,
         consume: {
@@ -92,6 +95,7 @@ async function syntheticActivityRoll(activity, targets = [], {options = {}, conf
         let spellLabel = actorUtils.getEquivalentSpellSlotName(activity.actor, atLevel);
         if (spellLabel) defaultConfig.spell = {slot: spellLabel};
     }
+    if (userId) options.asUser ||= userId;
     options = genericUtils.mergeObject(defaultOptions, options);
     config = genericUtils.mergeObject(defaultConfig, config);
     config.midiOptions = options;
@@ -143,7 +147,7 @@ async function syntheticItemDataRoll(itemData, actor, targets, {options = {}, co
 }
 async function syntheticActivityDataRoll(activityData, item, actor, targets, {options = {}, config = {}, atLevel = undefined, consumeUsage = false, consumeResources = false} = {}) {
     let itemData = genericUtils.duplicate(item.toObject());
-    itemData.system.activities[activityData.id] = activityData;
+    itemData.system.activities[activityData._id] = activityData;
     let newItem = await itemUtils.syntheticItem(itemData, actor);
     let newActivity = newItem.system.activities.get(activityData._id);
     return await syntheticActivityRoll(newActivity, targets, {options, config, atLevel, consumeUsage, consumeResources});
@@ -178,33 +182,26 @@ function preventDeath(ditem) {
     ditem.damageDetail.forEach(i => i.value = 0);
     ditem.damageDetail[0].value = ditem.totalDamage;
 }
-function modifyDamageAppliedFlat(ditem, modificationAmount) {
-    // We're gonna just assume this isn't healing, only damage
-    if (modificationAmount < 0) {
-        modificationAmount = Math.max(modificationAmount, -ditem.hpDamage - ditem.tempDamage);
-        // if (Math.abs(modificationAmount) > ditem.hpDamage) {
-        //     ditem.hpDamage = 0;
-        //     let tempMod = modificationAmount + ditem.hpDamage;
-        //     ditem.tempDamage += tempMod;
-        // }
-    // } else if (ditem.newTempHP) {
-    //     let tempMod = Math.max(0, ditem.newTempHP - modificationAmount);
-    //     let hpMod = -Math.min(0, ditem.newTempHP - modificationAmount);
-    //     ditem.tempDamage += tempMod;
-    //     ditem.hpDamage += hpMod;
-    // } else {
-    //     ditem.hpDamage += modificationAmount;
+function modifyDamageAppliedFlat(ditem, modificationAmount, {type = 'none', multiplier = 1} = {}) {
+    if (multiplier === 'auto') {
+        multiplier = ditem.damageDetail[0].active.multiplier;
+        let actor = fromUuidSync(ditem.actorUuid);
+        if (actor) {
+            if (actorUtils.checkTrait(actor, 'di', type)) modificationAmount = 0;
+            if (actorUtils.checkTrait(actor, 'dr', type)) {
+                modificationAmount = Math.floor(modificationAmount / 2);
+            }
+        }
     }
-    // ditem.hpDamage = Math.min(ditem.oldHP, ditem.damageDetail.reduce((acc, i) => acc + i.value, modificationAmount));
-    // ditem.hpDamage = Math.sign(ditem.hpDamage) * Math.floor(Math.abs(ditem.hpDamage));
+    if (modificationAmount < 0) modificationAmount = Math.max(modificationAmount, -ditem.hpDamage - ditem.tempDamage);
     ditem.damageDetail.push({
         value: modificationAmount,
-        active: {multiplier: 1},
-        type: 'none'
+        active: {multiplier},
+        type
     });
     ditem.rawDamageDetail.push({
         value: modificationAmount,
-        type: 'none'
+        type
     });
     let actualTotal = ditem.totalDamage + modificationAmount;
     ditem.totalDamage = actualTotal;