+++ scripts/macros.js
@@ -85,6 +85,7 @@ export {alterSelf} from './macros/2024/spells/alterSelf.js';
 export {animalFriendship} from './macros/2024/spells/animalFriendship.js';
 export {animalMessenger} from './macros/2024/spells/animalMessenger.js';
 export {animateDead} from './macros/2024/spells/animateDead.js';
+export {animateObjects} from './macros/2024/spells/animateObjects.js';
 export {antilifeShell, antilifeShellOtherMove} from './macros/2024/spells/antilifeShell.js';
 export {arcaneVigor} from './macros/2024/spells/arcaneVigor.js';
 export {armorOfAgathys} from './macros/2024/spells/armorOfAgathys.js';
@@ -146,7 +147,7 @@ export {fogCloud} from './macros/2024/spells/fogCloud.js';
 export {gaseousForm} from './macros/2024/spells/gaseousForm.js';
 export {globeOfInvulnerability, globeOfInvulnerabilityInvulnerable} from './macros/2024/spells/globeOfInvulnerability.js';
 export {grease} from './macros/2024/spells/grease.js';
-export {guardianOfFaith} from './macros/2024/spells/guardianOfFaith.js';
+export {guardianOfFaith, guardianOfFaithDamage} from './macros/2024/spells/guardianOfFaith.js';
 export {guidingBolt} from './macros/2024/spells/guidingBolt.js';
 export {gustOfWind, gustOfWindGust} from './macros/2024/spells/gustOfWind.js';
 export {heatMetal} from './macros/2024/spells/heatMetal.js';
@@ -261,7 +262,7 @@ export {sangromanticInitiate} from './macros/2024/feats/sangromanticInitiate.js'
 // Species Features
 // Aasimar
 export {healingHands} from './macros/2024/speciesFeatures/aasimar/healingHands.js';
-export {celestialRevelation, celestialRevelationInnerRadiance} from './macros/2024/speciesFeatures/aasimar/celestialRevelation.js';
+export {celestialRevelation, celestialRevelationInnerRadiance, celestialRevelationHeavenlyWings, celestialRevelationNecroticShroud} from './macros/2024/speciesFeatures/aasimar/celestialRevelation.js';
 export {celestialResistance} from './macros/2024/speciesFeatures/aasimar/celestialResistance.js';
 // Orc
 export {relentlessEndurance} from './macros/2024/speciesFeatures/orc/relentlessEndurance.js';
@@ -282,6 +283,9 @@ export {superiorAtlas, superiorAtlasEffect} from './macros/2024/classFeatures/ar
 export {ghaalShaarat, upgradeGhaalShaarat} from './macros/2024/classFeatures/artificer/forgeAdept/ghaalshaarat.js';
 export {perfectWeapon} from './macros/2024/classFeatures/artificer/forgeAdept/perfectWeapon.js';
 export {runesOfWar, runesOfWarAura, runesOfWarDamage} from './macros/2024/classFeatures/artificer/forgeAdept/runesOfWar.js';
+// Armorer
+export {arcaneArmor} from './macros/2024/classFeatures/artificer/armorer/arcaneArmor.js';
+export {armorModel, armorModelForceDemolisher, armorModelLightningLauncher, armorModelthunderPulseEffect} from './macros/2024/classFeatures/artificer/armorer/armorModel.js';
 // Barbarian
 export {rage, rageRaging, rageUpkeep} from './macros/2024/classFeatures/barbarian/rage.js';
 export {unarmoredDefenseBarbarian} from './macros/2024/classFeatures/barbarian/unarmoredDefense.js';
@@ -488,6 +492,7 @@ export {deflectEnergy} from './macros/2024/classFeatures/monk/deflectEnergy.js';
 export {disciplinedSurvivor} from './macros/2024/classFeatures/monk/disciplinedSurvivor.js';
 export {perfectFocus} from './macros/2024/classFeatures/monk/perfectFocus.js';
 export {superiorDefense} from './macros/2024/classFeatures/monk/superiorDefense.js';
+export {monkUnarmedStrike} from './macros/2024/classFeatures/monk/monkUnarmedStrike.js';
 // Warrior of the Elements
 export {elementalAttunement, elementalAttunementElementalStrikes} from './macros/2024/classFeatures/monk/warriorOfTheElements/elementalAttunement.js';
 export {elementalBurst} from './macros/2024/classFeatures/monk/warriorOfTheElements/elementalBurst.js';
@@ -523,7 +528,9 @@ export {schoolOfHardKnocks, schoolOfHardKnocksEndanger} from './macros/2024/clas
 export {shakeItOff} from './macros/2024/classFeatures/pugilist/shakeItOff.js';
 export {swaggerStreak} from './macros/2024/classFeatures/pugilist/swaggerStreak.js';
 // Squared Circle
+export {cleanFinish} from './macros/2024/classFeatures/pugilist/squaredCircle/cleanFinish.js';
 export {groundwork} from './macros/2024/classFeatures/pugilist/squaredCircle/groundwork.js';
+export {heavyweight} from './macros/2024/classFeatures/pugilist/squaredCircle/heavyweight.js';
 export {meatShield} from './macros/2024/classFeatures/pugilist/squaredCircle/meatShield.js';
 // Wizard
 export {arcaneRecovery} from './macros/2024/classFeatures/wizard/arcaneRecovery.js';